roll up mobile buildscript block to top level
While the buildscript block is only required for the mobile module,
keeping it in the top level results in gradle warning that the kotlin
plugin was loaded multiple times and could result in a build error. The
solution is to put plugin configuration at the top level, which in this
case is the project root.