clean up repository definitions (including minimizing jcenter usage)
Rolled up repo definitions in order to minimize the locations that they
are defined, namely top level and at the moblie root.

Also removed jcenter in all places possible, due to it shutting down in
May 2021. Mobile has a task that still depends on something in jcenter
so that is there for now, though hopefully can be removed soon.